mcp/pyproject.toml
[project]
name = "podcast-insight-mcp"
version = "0.1.0"
description = "MCP server for extracting insights from podcasts and videos"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "anthropic>=0.40.0",
    "yt-dlp>=2024.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
podcast-insight-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["server.py"]
